Extract resourses
