+++ 1_Converter/app/src/main/java/ppo/converter/ui/fragments/conversion/ConversionFragment.java
@@ -1,19 +1,14 @@
-package ppo.converter.ui.fragments;
+package ppo.converter.ui.fragments.conversion;
 
-import androidx.fragment.app.FragmentManager;
-import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 
-import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.ClipData;
 import android.content.ClipboardManager;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.content.res.Resources;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
