+++ 1_Converter/app/src/main/java/ppo/converter/ui/fragments/conversion/NumPadFragment.java
@@ -1,4 +1,4 @@
-package ppo.converter.ui.fragments;
+package ppo.converter.ui.fragments.conversion;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
