+++ 1_Converter/app/src/main/java/ppo/converter/viewModels/ModelFactory.java
@@ -20,9 +20,7 @@ public class ModelFactory extends ViewModelProvider.NewInstanceFactory {
     @NonNull
     @Override
     public <T extends ViewModel> T create(@NonNull Class<T> modelClass) {
-        if (modelClass == ConversionViewModel.class) {
-            return (T) new ConversionViewModel(res, title);
-        }
+        if (modelClass == ConversionViewModel.class) return (T) new ConversionViewModel(res, title);
         return null;
     }
 }
