+++ 1_Converter/app/src/main/res/layout/activity_main.xml
@@ -12,8 +12,8 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
-        android:layout_marginLeft="30dp"
-        android:layout_marginRight="30dp"
+        android:layout_marginLeft="@dimen/tabLayout_margin"
+        android:layout_marginRight="@dimen/tabLayout_margin"
         android:background="@color/colorPrimary"
         app:tabIndicatorColor="@color/colorPrimary"
         app:tabMode="fixed"
