--- 1_Converter/app/src/main/res/layout/fragment_conversion.xml
@@ -1,129 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="top"
-    android:paddingTop="30dp"
-    tools:context=".ui.fragments.ConversionFragment">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_marginLeft="30dp"
-        android:gravity="center"
-        android:layout_marginRight="30dp">
-
-        <LinearLayout
-            android:id="@+id/dimen1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:gravity="center">
-
-            <TextView
-                android:id="@+id/listbtn1"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:layout_marginRight="5dp"
-                android:textSize="25dp" />
-
-            <ImageView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:srcCompat="@android:drawable/arrow_down_float" />
-        </LinearLayout>
-
-        <Space
-            android:layout_width="35dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="0" />
-
-        <LinearLayout
-            android:id="@+id/number1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:gravity="right"
-            android:orientation="vertical">
-
-            <TextView
-                android:id="@+id/text1"
-                android:layout_width="250dp"
-                android:layout_height="wrap_content"
-                android:gravity="right"
-                android:text="1"
-                android:textSize="28dp" />
-
-            <TextView
-                android:id="@+id/unit1"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:textSize="12dp" />
-        </LinearLayout>
-
-    </LinearLayout>
-
-    <Space
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:layout_weight="0" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_marginLeft="30dp"
-        android:gravity="center"
-        android:layout_marginRight="30dp">
-
-        <LinearLayout
-            android:id="@+id/dimen2"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:gravity="center">
-
-            <TextView
-                android:id="@+id/listbtn2"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:layout_marginRight="5dp"
-                android:textSize="20dp" />
-
-            <ImageView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:srcCompat="@android:drawable/arrow_down_float" />
-        </LinearLayout>
-
-        <Space
-            android:layout_width="35dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="0" />
-
-        <LinearLayout
-            android:id="@+id/number2"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:gravity="right"
-            android:orientation="vertical">
-
-            <TextView
-                android:id="@+id/text2"
-                android:layout_width="250dp"
-                android:layout_height="wrap_content"
-                android:gravity="right"
-                android:enabled="false"
-                android:text="1"
-                android:textSize="28dp" />
-
-            <TextView
-                android:id="@+id/unit2"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:textSize="12dp"/>
-        </LinearLayout>
-    </LinearLayout>
-</LinearLayout>
