+++ 1_Converter/app/src/main/res/layout/fragment_length.xml
@@ -8,12 +8,9 @@
 
     <LinearLayout
         android:id="@+id/length_button"
-        android:layout_width="150dp"
-        android:layout_height="115dp"
-        android:layout_marginBottom="80dp"
+        style="@style/ButtonClickableStyle"
         android:background="@drawable/button_clicked_border"
-        android:gravity="center"
-        android:orientation="vertical">
+        android:layout_marginBottom="@dimen/length_fragment_margin_between_btns">
 
         <ImageView
             android:layout_width="54dp"
@@ -22,53 +19,46 @@
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Length"
-            android:textSize="20dp" />
+            android:text="@string/length"
+            android:textSize="@dimen/length_fragment_btn_text_size" />
 
     </LinearLayout>
 
     <LinearLayout
         android:id="@+id/area_button"
-        android:layout_width="150dp"
-        android:layout_height="115dp"
-        android:layout_marginBottom="80dp"
+        style="@style/ButtonClickableStyle"
         android:background="@drawable/button_clicked_border"
-        android:gravity="center"
-        android:orientation="vertical">
+        android:layout_marginBottom="@dimen/length_fragment_margin_between_btns">
 
         <ImageView
             android:layout_width="47dp"
             android:layout_height="45dp"
-            android:layout_marginBottom="10dp"
+            android:layout_marginBottom="@dimen/icon_text_margin"
             android:src="@drawable/ic_area" />
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Area"
-            android:textSize="20dp" />
+            android:text="@string/area"
+            android:textSize="@dimen/length_fragment_btn_text_size" />
 
     </LinearLayout>
 
     <LinearLayout
         android:id="@+id/volume_button"
-        android:layout_width="150dp"
-        android:layout_height="115dp"
         android:background="@drawable/button_clicked_border"
-        android:gravity="center"
-        android:orientation="vertical">
+        style="@style/ButtonClickableStyle">
 
         <ImageView
             android:layout_width="54dp"
             android:layout_height="46dp"
             android:layout_gravity="center"
-            android:layout_marginBottom="10dp"
-            android:background="@drawable/ic_volume"
-            android:elevation="2dp" />
+            android:layout_marginBottom="@dimen/icon_text_margin"
+            android:background="@drawable/ic_volume" />
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Volume"
-            android:textSize="20dp" />
+            android:text="@string/volume"
+            android:textSize="@dimen/length_fragment_btn_text_size"  />
 
     </LinearLayout>
