+++ 1_Converter/app/src/main/res/layout/list_item.xml
@@ -1,12 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<TextView android:id="@android:id/text1"
-    android:textSize="16sp"
+<TextView xmlns:tools="http://schemas.android.com/tools"
+    android:id="@android:id/text1"
+    android:textSize="@dimen/dialog_item_text_size"
     android:text="@string/app_name"
-    android:paddingTop="20dp"
-    android:paddingLeft="40dp"
-    android:paddingBottom="10dp"
+    android:paddingTop="@dimen/dialog_item_padding_top"
+    android:paddingStart="@dimen/dialog_item_padding_left"
+    android:paddingBottom="@dimen/dialog_item_padding_bottom"
     android:textColor="@color/colorText"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    xmlns:android="http://schemas.android.com/apk/res/android" />
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    tools:ignore="RtlSymmetry" />
