+++ 1_Converter/app/src/main/res/layout/num_pad_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragments.NumPadFragment"
+    tools:context=".ui.fragments.conversion.NumPadFragment"
     android:orientation="horizontal">
 
 
@@ -19,15 +19,15 @@
             <Button
                 style="@style/NumberStyle"
                 android:onClick="numOnClick"
-                android:text="7"/>
+                android:text="@string/_7"/>
             <Button
                 style="@style/NumberStyle"
                 android:onClick="numOnClick"
-                android:text="8"/>
+                android:text="@string/_8"/>
             <Button
                 style="@style/NumberStyle"
                 android:onClick="numOnClick"
-                android:text="9"/>
+                android:text="@string/_9"/>
         </TableRow>
 
         <TableRow
@@ -37,15 +37,15 @@
             <Button
                 style="@style/NumberStyle"
                 android:onClick="numOnClick"
-                android:text="4" />
+                android:text="@string/_4" />
             <Button
                 style="@style/NumberStyle"
                 android:onClick="numOnClick"
-                android:text="5"/>
+                android:text="@string/_5"/>
             <Button
                 style="@style/NumberStyle"
                 android:onClick="numOnClick"
-                android:text="6"/>
+                android:text="@string/_6"/>
         </TableRow>
 
         <TableRow
@@ -55,15 +55,15 @@
             <Button
                 style="@style/NumberStyle"
                 android:onClick="numOnClick"
-                android:text="1"/>
+                android:text="@string/_1"/>
             <Button
                 style="@style/NumberStyle"
                 android:onClick="numOnClick"
-                android:text="2"/>
+                android:text="@string/_2"/>
             <Button
                 style="@style/NumberStyle"
                 android:onClick="numOnClick"
-                android:text="3"/>
+                android:text="@string/_3"/>
         </TableRow>
 
         <TableRow
@@ -77,11 +77,11 @@
             <Button
                 style="@style/NumberStyle"
                 android:onClick="numOnClick"
-                android:text="0"/>
+                android:text="@string/_0"/>
             <Button
                 style="@style/NumberStyle"
                 android:onClick="numOnClick"
-                android:text="."/>
+                android:text="@string/dot"/>
         </TableRow>
     </TableLayout>
 
@@ -99,8 +99,8 @@
             android:layout_weight="1"
             android:background="@drawable/delete_clicked_border"
             android:onClick="AC"
-            android:text="AC"
-            android:textSize="24dp"
+            android:text="@string/ac"
+            android:textSize="@dimen/ac_text_size"
             android:textColor="@color/colorAccent">
         </Button>
